update make it readable

+++ app/resources/answers/answers_views.py
@@ -1,5 +1,5 @@
 '''this module is intented to do all the routing in question api'''
-import datetime
+from datetime import datetime
 import time
 import psycopg2
 import psycopg2.extras
@@ -16,28 +16,48 @@ from . import ANSWERS
 from flask_jwt_extended import jwt_required,get_jwt_identity
 
 answers=Answers()
-queston=Questions()
+question=Questions()
 @ANSWERS.route("/api/v1/answers/<int:questionid>", methods=["GET","POST"])
 @jwt_required
 def question_view(questionid):
     """show all answer of a question"""
+    current_user=get_jwt_identity()
     try:
         connection=database_connection("development")
+
         try:
-            cursor = answers.search_answer_by_questionid(questionid, connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-            answer = cursor.fetchall()
-            if answer:
-                quiz=queston.search_question_by_questionid(questionid,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-                result={}
-                result["question"]=quiz[0]
-                result["answers"]=answers
-                return jsonify({"answers": result}),200
-            abort(404)
+            cursor=question.search_question_by_questionid(questionid,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+            quiz=cursor.fetchall()
+            if quiz:
+                if request.method=="GET":
+                    result={}
+                    result["question"]=quiz[0]
+                    result["answers"]=answers
+                    return jsonify({"answers": result}),200
+                if not request.json:
+                    return jsonify({"error":"empty body"})
+                if not "answer_text" in request.json:
+                    return jsonify({"error":"answer_text key not found"})
+                answer_text=request.json["answer_text"]
+                if answer_text.strip():
+                    time_created=datetime.utcnow()
+                    vote=0
+                    is_answer=False
+                    userid=current_user[0]["userid"]
+                    answers.add_answer(answer_text,time_created,userid,questionid,vote,is_answer,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+                    cursor = answers.search_answer_by_questionid(questionid, connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+                    answers_list = cursor.fetchall()
+                    result={}
+                    result["question"]=quiz[0]
+                    result["answers"]=answers_list
+                    return jsonify({"answers": result}),200
+                return jsonify({"error":"answer cannot be empty"}),400
+            return jsonify({"error":"question you are trying to answer is not available"}),404
         except (Exception, psycopg2.DatabaseError) as error:
-            return jsonify({"error":str(error)}),500
+            return jsonify({"error":str(error)})#jsonify({"error": "request error please check your request body"}),400
         connection.close()
     except (Exception, psycopg2.DatabaseError) as e:
-        return jsonify({"error":str(e)}),500
+        return jsonify({"error": "request error please check your request body"}),400
 
 @ANSWERS.route('/api/v1/update_answer/<int:answerid>', methods=["PUT"])
 @jwt_required
@@ -53,12 +73,16 @@ def update_answer(answerid):
             cursor=answers.search_answer_by_id(answerid,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
             answer_list=cursor.fetchall()
             if answer_list:
-                answers.update_answer(answer_text,answerid,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-                return jsonify({"info":"answer updated"}),200
+                answer=answer_list[0]
+                question_cursor=question.search_question_by_questionid(answer["questionid"],connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+                questions_list=question_cursor.fetchall()
+                if current_user[0]["userid"]==questions_list[0]["userid"]:
+                    answers.update_answer(answer_text,answerid,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+                    return jsonify({"info":"answer updated"}),200
         except Exception as error:
-            return jsonify({"error": str(error)}),500
+            return jsonify({"error": "request error please check your request body"}),400
     except (Exception, psycopg2.DatabaseError) as e:
-        return jsonify({"error":str(e)}),500
+        return jsonify({"error": "request error please check your request body"}),400
 
 @ANSWERS.route('/api/v1/delete_answer/<int:answerid>', methods=["DELETE"])
 @jwt_required
@@ -69,36 +93,41 @@ def delete_answer(answerid):
         connection=database_connection("development")
         try:
             cursor=answers.search_answer_by_id(answerid, connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-            answers.delete_answer(answerid,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-            return jsonify({"info":"deleted"}),200
+            answers_list=cursor.fetchall()
+            if answers_list:
+                answer=answers_list[0]
+                question_cursor=question.search_question_by_questionid(answer["questionid"],connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+                questions_list=question_cursor.fetchall()
+                if current_user[0]["userid"]==questions_list[0]["userid"]:
+                    answers.delete_answer(answerid,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+                    return jsonify({"info":"deleted"}),200
+                abort(403)
         except (Exception, psycopg2.DatabaseError) as error:
-            return jsonify({"error":str(error)}),500
+            return jsonify({"error": "request error please check your request body"}),400
     except (Exception, psycopg2.DatabaseError) as e:
-        return jsonify({"error":str(e)}),500
+        return jsonify({"error": "request error please check your request body"}),400
 
 @ANSWERS.route('/api/v1/prefered_answer/<int:answerid>', methods=["PATCH"])
 @jwt_required
 def mark_prefered(answerid):
     """ mark answer as prefered"""
     current_user = get_jwt_identity()
+    connection = database_connection("development")
     try:
-        connection=database_connection("development")
-        if not request.json:
-            abort(400)
-        if not "is_answer" in request.json:
-            abort(400)
-        try:
-            cursor=answers.search_answer_by_id(answerid, connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-            answer_list=cursor.fetchall()
-            if answer_list:
+        cursor=answers.search_answer_by_id(answerid, connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+        answer_list=cursor.fetchall()
+        if answer_list:
+            questionid=answer_list[0]["questionid"]
+            question_cursor=question.search_question_by_questionid(questionid,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+            questions_list=question_cursor.fetchall()
+            if current_user[0]["userid"]==questions_list[0]["userid"]:
                 answers.mark_prefered(answerid, True, connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-                answer_cursor = answer.search_answer_by_questionid(question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+                answer_cursor = answers.search_answer_by_questionid(questionid,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
                 question_answer=answer_cursor.fetchall()
-                question_cursor=question.search_question_by_questionid(question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+                question_cursor=question.search_question_by_questionid(questionid,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
                 new_question=question_cursor.fetchall()
                 return jsonify({"question":new_question[0],"answers":question_answer}),200
-            abort(404)
-        except (Exception,psycopg2.DatabaseError) as error:
-            return jsonify({"error": str(error)})
+            abort(403)
+        return jsonify({"error":"answer not found"})
     except (Exception, psycopg2.DatabaseError) as e:
-        return jsonify({"error":str(e)}),500
+        return jsonify({"error":str(e)})#jsonify({"error": "request error please check your request body"}),400
