+++ app/resources/questions/questions.py
@@ -4,7 +4,6 @@ import time
 import psycopg2
 from flask import redirect
 from flask import url_for
-import re
 from flask import request
 from flask import jsonify
 from flask import abort
@@ -32,52 +31,30 @@ def home():
                 return jsonify({"questions": questions_list}),200
             abort(404)
         except (Exception, psycopg2.DatabaseError) as error:
-            return jsonify({"error":str(error)}),500
+            return jsonify({"error":str(error)}),400
         connection.close()
     except (Exception, psycopg2.DatabaseError) as e:
-        return jsonify({"error":str(e)}),500
+        return jsonify({"error": "request error please check your request body"}),400
 
-@QUESTION.route('/api/v1/questions/<int:question_id>', methods=["GET","POST"])
+@QUESTION.route('/api/v1/questions/<int:question_id>', methods=["GET"])
 @jwt_required
 def get_question(question_id):
     '''get a specific question'''
     current_user = get_jwt_identity()
-    if request.method == 'GET':
+    try:
+        connection=database_connection("development")
         try:
-            connection=database_connection("development")
-            try:
-                cursor=question.search_question_by_questionid(question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-                new_question=cursor.fetchall()
+            cursor=question.search_question_by_questionid(question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+            new_question=cursor.fetchall()
+            if new_question:
                 return jsonify({"question":new_question[0]}),200
-            except (Exception, psycopg2.DatabaseError) as error:
-                return jsonify({"error":str(error)}),500
-            abort(404)
-            connection.close()
-        except Exception as e:
-            return jsonify({"error":"database connection error"}),500
-    elif request.method == "POST":
-        try:
-            connection=database_connection("development")
-            if not request.json or not "answer_text" in request.json:
-                abort(400)
-            if not "userid" in request.json:
-                abort(400)
-            now=time.time()
-            time_created=now
-            answer_text=request.json["answer_text"]
-            userid=request.json["userid"]
-            try:
-                answer.add_answer(answer_text,time_created,userid,question_id,0,False,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-                answer_cursor = answer.search_answer_by_questionid(question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-                question_answer=answer_cursor.fetchall()
-                question_cursor=question.search_question_by_questionid(question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-                new_question=question_cursor.fetchall()
-                return jsonify({"question":new_question[0],"answers":question_answer}),200
-            except (Exception,psycopg2.DatabaseError) as error:
-                return jsonify({"error":str(error)}),500
-            connection.close()
-        except (Exception, psycopg2.DatabaseError) as e:
-            return jsonify({"error":str(e)}),500
+            return jsonify({"error":"question not found"})
+        except (Exception, psycopg2.DatabaseError) as error:
+            return jsonify({"error":str(error)}),400
+        abort(404)
+        connection.close()
+    except Exception as e:
+        return jsonify({"error": "request error please check your request body"}),400
 
 @QUESTION.route('/api/v1/add_question', methods=["POST"])
 @jwt_required
@@ -93,16 +70,19 @@ def post_question():
         if not "description" in request.json:
             return jsonify({"error":"key error 'description' in your request"})
         userid=current_user[0]["userid"]
-        title= request.json['title'],
-        description = request.json['description'],
+        title= request.json['title']
+        description = request.json['description']
         time_created = datetime.utcnow()
         try:
-            question.add_question(title,description,time_created,userid,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-            return jsonify({"info":"added"}),201
+            if title.strip() and description.strip():
+                question.add_question(title,description,time_created,userid,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+                return jsonify({"info":"question asked"}),201
+            return jsonify({"error":"you have an empty key value in you body"})
         except (Exception, psycopg2.DatabaseError) as error:
-            return jsonify({"error":str(error)}),500
+            return jsonify({"error":str(error)}) #jsonify({"error": "request error please check your request body"}),400
+
     except (Exception, psycopg2.DatabaseError) as e:
-        return jsonify({"error":str(e) + "not found"}),500
+        return jsonify({"error": "request error please check your request body"}),400
 
 @QUESTION.route('/api/v1/update_question/<int:question_id>', methods=["PUT"])
 @jwt_required
@@ -121,17 +101,21 @@ def update_question(question_id):
             cursor=question.search_question_by_questionid(question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
             current_question=cursor.fetchall()
             if current_question:
-                title = request.json.get('title')
-                description= request.json.get('description')
-                question.update_question(title, description, question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-                cursor=question.search_question_by_questionid(question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-                new_question=cursor.fetchall()
-                return jsonify({"question":new_question}),200
-            abort(404)
+                if current_user[0]["userid"]==current_question[0]["userid"]:
+                    title = request.json.get('title')
+                    description= request.json.get('description')
+                    if title.strip() and description.strip():
+                        question.update_question(title, description, question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+                        cursor=question.search_question_by_questionid(question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+                        new_question=cursor.fetchall()
+                        return jsonify({"question":new_question}),200
+                    return jsonify({"error":"you have an empty key value in you body"}),400
+                abort(403)
+            return jsonify({"error":"no question found"}),404
         except Exception as error:
-            return jsonify({"error":str(error)}),500
+            return jsonify({"error": "request error please check your request body"}),400
     except (Exception, psycopg2.DatabaseError) as e:
-        return jsonify({"error":str(e)}),500
+        return jsonify({"request error":"request error please check your request body"}),400
 
 @QUESTION.route('/api/v1/delete_question/<int:question_id>', methods=["DELETE"])
 @jwt_required
@@ -144,16 +128,20 @@ def delete_question(question_id):
             cursor=question.search_question_by_questionid(question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
             delete_question=cursor.fetchall()
             if delete_question:
-                if question.delete_question(question_id,connection.cursor()):
-                    return jsonify({"result": "deleted"}),200
-            return jsonify({"error": "no question found"})
+                if current_user[0]["userid"] == delete_question[0]["userid"]:
+                    if question.delete_question(question_id,connection.cursor()):
+                        return jsonify({"result": " question deleted"}),200
+                    return jsonify({"error": "no question found"})
+                abort(403)
+            return jsonify({"error":"no question found"}),404
         except (Exception,psycopg2.DatabaseError) as error:
-            return jsonify({"error": str(error)}),500
+            return jsonify({"error": "request error please check your request body"}),400
         connection.close()
     except (Exception, psycopg2.DatabaseError) as e:
-        return jsonify({"error":str(e)}),500
+        return jsonify({"error": "request error please check your request body"}),400
 
 @QUESTION.route('/api/v1/search', methods=["POST"])
+@jwt_required
 def search_question():
     """search question question"""
     current_user = get_jwt_identity()
@@ -170,9 +158,9 @@ def search_question():
             questions = cursor.fetchall()
             if questions:
                     return jsonify({"questions": questions}),200
-            return jsonify({"error": "no question found"})
+            return jsonify({"error": "no question found"}),404
         except (Exception,psycopg2.DatabaseError) as error:
             return jsonify({"error": str(error)})
         connection.close()
     except (Exception, psycopg2.DatabaseError) as e:
-        return jsonify({"error":str(e)}),500
+        return jsonify({"error":str(e)}),400
